+++ Sources/RSocketCore/Extensions/Coder/Decoder+CompositeMetadata.swift
@@ -1,0 +1,63 @@
+
+/*
+ * Copyright 2015-present the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import Foundation
+
+extension Decoders {
+    public struct CompositeMetadataDecoder<Decoder, CompositeMetadataDecoder>: DecoderProtocol where
+    Decoder: DecoderProtocol,
+    Decoder.Metadata == [CompositeMetadata],
+    CompositeMetadataDecoder: RSocketCore.CompositeMetadataDecoder
+    {
+        public typealias Metadata = CompositeMetadataDecoder.Metadata
+        public typealias Data = Decoder.Data
+        
+        @usableFromInline
+        internal var decoder: Decoder
+        
+        @usableFromInline
+        internal let metadataDecoder: CompositeMetadataDecoder
+        
+        @usableFromInline
+        internal init(
+            decoder: Decoder, 
+            metadataDecoder: CompositeMetadataDecoder
+        ) {
+            self.decoder = decoder
+            self.metadataDecoder = metadataDecoder
+        }
+        
+        @inlinable
+        mutating public func decode(
+            _ payload: Payload,
+            mimeType: ConnectionMIMEType
+        ) throws -> (Metadata, Data) {
+            let (metadata, data) = try decoder.decode(payload, mimeType: mimeType)
+            let decodedMetadata = try metadataDecoder.decode(from: metadata)
+            return (decodedMetadata, data)
+        }
+    }
+}
+
+extension DecoderProtocol {
+    @inlinable
+    public func decodeMetadata<CompositeMetadataDecoder>(
+        @CompositeMetadataDecoderBuilder metadataDecoder: () -> CompositeMetadataDecoder
+    ) -> Decoders.CompositeMetadataDecoder<Self, CompositeMetadataDecoder> {
+        .init(decoder: self, metadataDecoder: metadataDecoder())
+    }
+}
